chore: rename AddTokenModal to AddToken
