+++ apps/mobile-wallet/src/components/AddToken.vue
